requirements.txt
asn1crypto==0.24.0
boto==2.48.0
boto3==1.5.22
botocore==1.8.36
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
dj-database-url==0.4.2
Django==1.10
django-filter==1.1.0
django-storages==1.6.5
djangorestframework==3.7.7
docutils==0.14
enum34==1.1.6
futures==3.2.0
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.19
jmespath==0.9.3
Markdown==2.6.11
Pillow==5.0.0
psycopg2==2.7.3.2
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.5.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
twilio==6.10.1
urllib3==1.22
whitenoise==3.3.1
